Put the extra hypothesis flags in the root level conftest.py
pytest_addoption() doesn't work in package-level conftest.py files. These
aren't critical for running the tests, so it isn't a big deal to have them in
a separate conftest that isn't installed.
